requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pika==1.3.2
gunicorn==21.2.0